Check all term arguments for freshness in injectivity
Currently Tamarin will only check the first term argument for any
given injective fact to test whether it is fresh. However, the
definition of injectivity allows any term argument to be the fresh
term, as long as it is consistent across all rules.

This is my first attempt at doing that within Tamarin. It works by
enumerating all term arguments when looking for candidate facts and
then filtering out those that are never fresh in the constructor or
are not consistently kept in the same position by the other rules.