First commit for graph execution plotting.
Given a theory file, attempt to produce a dot graph visualising how the
protocol will execute. It does this by taking the protocol rules and
attempting to unify each of their premises against the conclusions of
all of the protocol rules. If it can successfully unify, then we add an
edge to the graph. It can handle persistent facts, as well as In/Out
pairs (i.e. it ignores the presence of an attacker).

Currently this feature can only be called from the command line
(--graph) and produces a dot file, which can then be compiled into an
image file. However, this may be useful for the web interface too.