Create a single file containing all arguments
In order to display all the arguments for tamarin in a single page, we
need to be able to reference them all. Unfortunately doing this within
Console.hs introduces circular dependencies. Instead, I have created a
new file src/Main/Flags.hs, which will list all of the different flags
needed for each mode and export them as necessary.

This is still work in progress - the unified help output needs to group
the flags by mode.